add tests for remote schema subscription
