+++ server/tests-py/test_schema_stitching.py
@@ -1,16 +1,12 @@
-#!/usr/bin/env python3
-
-import time
-import yaml
 import json
-import queue
-from urllib.parse import urlparse
+import string
+import random
 
+import yaml
 import requests
 import pytest
 
 from validate import check_query_f, check_query
-from ws_graphql_client import GraphQLClient
 
 
 def mk_add_remote_q(name, url):
@@ -258,15 +254,35 @@ class TestRemoteSchemaQueriesOverWebsocket:
         assert st_code == 200, resp
 
     def _init(self, hge_ctx):
-        hge_ctx.ws.send(json.dumps({'type': 'connection_init', 'payload': {}}))
+        payload = {'type': 'connection_init', 'payload': {}}
+        if hge_ctx.hge_key:
+            payload['payload']['headers'] = {
+                'x-hasura-admin-secret': hge_ctx.hge_key
+            }
+        hge_ctx.ws.send(json.dumps(payload))
         ev = hge_ctx.get_ws_event(3)
         assert ev['type'] == 'connection_ack', ev
 
-    def _stop(self, hge_ctx, id):
-        data = {'id': id, 'type': 'stop'}
+    def _stop(self, hge_ctx, _id):
+        data = {'id': _id, 'type': 'stop'}
         hge_ctx.ws.send(json.dumps(data))
         ev = hge_ctx.get_ws_event(3)
         assert ev['type'] == 'complete', ev
+
+    def _send_query(self, hge_ctx, query):
+        self._init(hge_ctx)
+        _id = gen_id()
+        frame = {
+            'id': _id,
+            'type': 'start',
+            'payload': {'query': query},
+        }
+        if hge_ctx.hge_key:
+            frame['payload']['headers'] = {
+                'x-hasura-admin-secret': hge_ctx.hge_key
+            }
+        hge_ctx.ws.send(json.dumps(frame))
+        return _id
 
     def test_remote_query(self, hge_ctx):
         self._init(hge_ctx)
@@ -278,16 +294,11 @@ class TestRemoteSchemaQueriesOverWebsocket:
           }
         }
         """
-        frame = {
-            'id': '123',
-            'type': 'start',
-            'payload': {'query': query},
-        }
-        hge_ctx.ws.send(json.dumps(frame))
+        _id = self._send_query(hge_ctx, query)
         ev = hge_ctx.get_ws_event(3)
-        assert ev['type'] == 'data' and ev['id'] == '123', ev
+        assert ev['type'] == 'data' and ev['id'] == _id, ev
         assert ev['payload']['data']['data']['user']['username'] == 'john'
-        self._stop(hge_ctx, '123')
+        self._stop(hge_ctx, _id)
 
     def test_remote_mutation(self, hge_ctx):
         self._init(hge_ctx)
@@ -301,117 +312,13 @@ class TestRemoteSchemaQueriesOverWebsocket:
           }
         }
         """
-        frame = {
-            'id': '124',
-            'type': 'start',
-            'payload': {'query': query},
-        }
-        hge_ctx.ws.send(json.dumps(frame))
+        _id = self._send_query(hge_ctx, query)
         ev = hge_ctx.get_ws_event(3)
-        assert ev['type'] == 'data' and ev['id'] == '124', ev
+        assert ev['type'] == 'data' and ev['id'] == _id, ev
         assert ev['payload']['data']['data']['createUser']['user']['id'] == 42
         assert ev['payload']['data']['data']['createUser']['user']['username'] == 'foobar'
-        self._stop(hge_ctx, '124')
+        self._stop(hge_ctx, _id)
 
-
-@pytest.mark.skip(reason='does not have proper env')
-class TestRemoteSchemaSubscriptions():
-    dir = 'queries/remote_schemas'
-    #remote_url = 'ws://localhost:8080/v1alpha1/graphql'
-    teardown = {"type": "clear_metadata", "args": {}}
-
-    @pytest.fixture(autouse=True)
-    def transact(self, hge_ctx):
-        q = mk_add_remote_q('subs-test',
-                            'http://localhost:8081/v1alpha1/graphql')
-        st_code, resp = hge_ctx.v1q(q)
-        assert st_code == 200, resp
-        yield
-        # teardown
-        st_code, resp = hge_ctx.v1q(self.teardown)
-        assert st_code == 200, resp
-
-    def _mk_ws_url(self, url):
-        ws_url = urlparse(url)._replace(scheme='ws', path='/v1alpha1/graphql')
-        return ws_url.geturl()
-
-    def _run_subscription(self, hge_ctx, ws, query, cb, subalive=3):
-        headers = {}
-        if hge_ctx.hge_key:
-            headers = {'x-hasura-admin-secret': hge_ctx.hge_key}
-
-        id = ws.subscribe(query, callback=cb)
-        time.sleep(subalive)
-        ws.stop_subscribe(id)
-
-    def test_simple_subscription(self, hge_ctx):
-        query = "subscription { animals {id common_name} }"
-        def cb(_id, data):
-            assert len(data['payload']['data']['data']['animals']) == 3
-
-        ws = GraphQLClient(self._mk_ws_url(hge_ctx.hge_url))
-        self._run_subscription(hge_ctx, ws, query, cb)
-        ws.close()
-
-    def test_remote_then_local_subscription(self, hge_ctx):
-        q1 = "subscription { animals {id common_name} }"
-        q2 = "subscription { hello { code name } }"
-        def animals_sub(_id, data):
-            assert len(data['payload']['data']['data']['animals']) == 3
-
-        def hello_sub(_id, data):
-            assert len(data['payload']['data']['data']['code']) == 2
-
-        ws = GraphQLClient(self._mk_ws_url(hge_ctx.hge_url))
-        self._run_subscription(hge_ctx, ws, q1, animals_sub)
-        time.sleep(1)
-        self._run_subscription(hge_ctx, ws, q2, hello_sub)
-        ws.close()
-
-    def test_local_then_remote_subscription(self, hge_ctx):
-        q1 = "subscription { hello { code name } }"
-        q2 = "subscription { animals {id common_name} }"
-        def animals_sub(_id, data):
-            assert len(data['payload']['data']['data']['animals']) == 3
-
-        def hello_sub(_id, data):
-            assert len(data['payload']['data']['data']['code']) == 2
-
-        ws = GraphQLClient(self._mk_ws_url(hge_ctx.hge_url))
-        self._run_subscription(hge_ctx, ws, q1, hello_sub)
-        time.sleep(1)
-        self._run_subscription(hge_ctx, ws, q2, animals_sub)
-        ws.close()
-
-    def test_consecutive_remote_subscription(self, hge_ctx):
-        q = "subscription { animals { id common_name } }"
-        def animals_sub(_id, data):
-            assert len(data['payload']['data']['data']['animals']) == 3
-
-        ws = GraphQLClient(self._mk_ws_url(hge_ctx.hge_url))
-        self._run_subscription(hge_ctx, ws, q, animals_sub)
-        time.sleep(1)
-        self._run_subscription(hge_ctx, ws, q, animals_sub)
-        ws.close()
-
-    def test_interleaved_remote_subscription(self, hge_ctx):
-        q1 = "subscription { animals { id common_name } }"
-        q2 = "subscription { languages { name type } }"
-        def animals_sub(_id, data):
-            assert len(data['payload']['data']['data']['animals']) == 3
-
-        def lang_sub(_id, data):
-            assert len(data['payload']['data']['data']['languages']) == 4
-
-        ws = GraphQLClient(self._mk_ws_url(hge_ctx.hge_url))
-        id1 = ws.subscribe(q1, callback=animals_sub)
-        time.sleep(1)
-        id2 = ws.subscribe(q2, callback=lang_sub)
-        time.sleep(2)
-        ws.stop_subscribe(id1)
-        ws.stop_subscribe(id2)
-        time.sleep(1)
-        ws.close()
 
 
 class TestAddRemoteSchemaCompareRootQueryFields:
@@ -510,3 +417,6 @@ def check_introspection_result(res, types, node_names):
             satisfy_node = False
 
     return satisfy_node and satisfy_ty
+
+def gen_id(size=6, chars=string.ascii_letters + string.digits):
+    return ''.join(random.choice(chars) for _ in range(size))
